+++ client/src/pages/SoilGame/__tests__/testGameHelper.test.ts
@@ -37,9 +37,9 @@ import {
   formatQuestProgress,
   getElementSymbol,
   formatLocationInfo,
-  isValidCommand,
-  parseCommand
 } from '../utils/QuestListHelper';
+
+import { isValidCommand, parseCommand } from '../utils/CommandParser';
 
 import {
   isValidPosition,
